# internal/domain/prices.go
package domain

type Price struct {
	ID       string   `json:"id"`
	ItemId   string   `json:"item_id"`
	Amount   float64  `json:"amount"`
	Currency Currency `json:"currency"`
}

type Currency struct {
	ID               string `json:"id"`
	Symbol           string `json:"symbol"`
	DecimalDivider   string `json:"decimal_divider"`
	ThousandsDivider string `json:"thousands_divider"`
}

type Prices struct {
	Prices []Price `json:"prices"`
}